refactor(asset): moving classification definition to schema
