+++ modules/classifications/auxiliaries/define/keeper.go
@@ -12,6 +12,7 @@ import (
 	errorConstants "github.com/AssetMantle/modules/schema/errors/constants"
 	"github.com/AssetMantle/modules/schema/helpers"
 	baseIDs "github.com/AssetMantle/modules/schema/ids/base"
+	"github.com/AssetMantle/modules/schema/mappables/base"
 	"github.com/AssetMantle/modules/utilities/property"
 )
 
@@ -39,7 +40,7 @@ func (auxiliaryKeeper auxiliaryKeeper) Help(context sdkTypes.Context, request he
 		return newAuxiliaryResponse(classificationID, errorConstants.EntityAlreadyExists)
 	}
 
-	classifications.Add(mappable.NewClassification(auxiliaryRequest.Immutables, auxiliaryRequest.Mutables))
+	classifications.Add(mappable.NewMappable(base.NewClassification(auxiliaryRequest.Immutables, auxiliaryRequest.Mutables)))
 
 	return newAuxiliaryResponse(classificationID, nil)
 }
