+++ modules/classifications/internal/mappable/mappable_test.go
@@ -23,7 +23,7 @@ func Test_Classification_Methods(t *testing.T) {
 	id := baseIDs.NewClassificationID(immutables, mutables)
 
 	testClassification := NewClassification(immutables, mutables)
-	require.Equal(t, classification{Document: baseQualified.NewDocument(id, immutables, mutables)}, testClassification)
+	require.Equal(t, mappable{Document: baseQualified.NewDocument(id, immutables, mutables)}, testClassification)
 	require.Equal(t, immutables, testClassification.GetImmutables())
 	require.Equal(t, mutables, testClassification.GetMutables())
 }
